Add class Diagram and change the directory
