+++ two-step-view/src/main/java/com/iluwater/twostepview/XsltTransformer.java
@@ -22,8 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package iluwater.twostepview;
-
+package com.iluwater.twostepview;
 
 import javax.xml.transform.*;
 import javax.xml.transform.stream.*;
@@ -31,22 +30,23 @@ import java.io.*;
 
 
 /**
- * In this class,it is the second stage of the two-step view pattern which produced final HTML page
- * It transforms the input HTML file with XSLT stylesheet (toHTML.xsl) to describe
+ * In this class,it is the first stage of the two-step view pattern which produced page in logical representation.
+ * It transforms the input XML document (couse.xml) with XSLT stylesheet (coursetostudent.xsl) to describe
+ * how the course data is transformed into another XML document (students.xml) that uses the formatting of
+ * remove the course title and map each student's personal details (userID, firstName, lastName and DOB).
  */
-
-public class HtmlTransformer {
+public class XsltTransformer {
 
     /**
-     * @author Fei Lin
-     * @param args transform XML to HTML with XSLT
+     * @author Stephanie Sun
+     * @param args transform XML to another XML with XSLT
      */
     public static void main(String[] args) {
 
         // Files that requires for the transformation process
-        String xslFile ="two-step-view/src/main/xml.xslt/tohtml.xsl";
-        String inputFile = "two-step-view/src/main/xml.xslt/students.xml";
-        String outputFile = "two-step-view/src/main/xml.xslt/students.html";
+        String xslFile ="two-step-view/src/main/xml.xslt/coursetostudent.xsl";
+        String inputFile = "two-step-view/src/main/xml.xslt/course.xml";
+        String outputFile = "two-step-view/src/main/xml.xslt/students.xml";
 
         StreamSource xsl = new StreamSource(new File(xslFile));
         StreamSource input = new StreamSource(new File(inputFile));
@@ -62,5 +62,6 @@ public class HtmlTransformer {
         } catch (TransformerException e){
             e.printStackTrace();
         }
+
     }
 }
