+++ two-step-view/src/test/java/com/iluwatar/twostepview/XsltTransformerTest.java
@@ -1,6 +1,6 @@
 package com.iluwatar.twostepview;
 
-import iluwater.twostepview.XsltTransformer;
+import com.iluwater.twostepview.XsltTransformer;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
